实时行情价+K线图

+++ YiBei/app/src/main/java/com/ts/lys/yibei/customeview/ChooseTimesLayout.java
@@ -35,7 +35,7 @@ public class ChooseTimesLayout extends AutoLinearLayout implements View.OnClickL
     /**
      * 最低购买手数
      */
-    private float mintimes = 0.01f;
+    private double mintimes = 0.01;
 
 
     public ChooseTimesLayout(Context context) {
@@ -81,7 +81,7 @@ public class ChooseTimesLayout extends AutoLinearLayout implements View.OnClickL
     /**
      * 最低购买手数
      */
-    public void setBaseTimes(float minTimes) {
+    public void setBaseTimes(double minTimes) {
         this.mintimes = minTimes;
     }
